refactor(domain): LottoWinningType에서의 당첨금의 타입을 int에서 long으로 수정
< description >
- 기존의 int일땐 수익률이 마이너스로 찍히는 문제가 발생하였었음.

+++ src/main/java/lotto/domain/LottoWinningResult.java
@@ -13,7 +13,7 @@ public class LottoWinningResult {
 
     public long calculateTotalWinningAmount() {
         return allTypesCount.keySet().stream()
-                .mapToInt(type -> type.getWinningAmount() * allTypesCount.get(type))
+                .mapToLong(type -> type.getWinningAmount() * allTypesCount.get(type))
                 .sum();
     }
